+++ src/noahtheduke/splint/config.clj
@@ -6,9 +6,13 @@
   (:require
     [clojure.edn :as edn]
     [clojure.java.io :as io]
-    [clojure.set :as set]))
+    [clojure.set :as set]
+    [clojure.string :as str]))
 
 (set! *warn-on-reflection* true)
+
+(def version (str/trim (slurp "./resources/SPLINT_VERSION")))
+(def splint-version (str "splint v" version))
 
 (defn read-default-config []
   (edn/read-string (slurp (io/resource "config/default.edn"))))
@@ -20,11 +24,13 @@
   (loop [dir (.getParentFile (.getAbsoluteFile (io/file ".")))]
     (let [config (io/file dir ".splint.edn")]
       (if (.exists config)
-        (edn/read-string (slurp config))
+        {:dir dir
+         :file (.getAbsoluteFile config)
+         :local (edn/read-string (slurp config))}
         (when-let [parent (.getParentFile dir)]
           (recur parent))))))
 
-(defn deep-merge [& maps]
+(defn deep-merge [default & maps]
   (letfn [(reconcile-keys [val-in-result val-in-latter]
             (if (and (map? val-in-result)
                      (map? val-in-latter))
@@ -32,16 +38,16 @@
               val-in-latter))
           (reconcile-maps [result latter]
             (merge-with reconcile-keys result latter))]
-    (reduce reconcile-maps {} maps)))
+    (reduce reconcile-maps default maps)))
 
-(defn load-config [options]
-  (let [local (find-local-config)
-        merged (deep-merge @default-config local)]
-    (-> merged
-        (set/rename-keys {'output :output
-                          'parallel :parallel
-                          'quiet :quiet})
-        (merge options))))
+(defn load-config
+  ([options] (load-config (:local (find-local-config)) options))
+  ([local options]
+   (-> (deep-merge @default-config local)
+       (set/rename-keys {'output :output
+                         'parallel :parallel
+                         'quiet :quiet})
+       (merge options))))
 
 (defn get-config [ctx rule]
   (let [full-name (:full-name rule)